2.2.0 - ID change + new components
- Mod ID has been updated to fifymcc (Breaking change)
- Enchantability component has been added and hooked up
- FuelValue component has been added and hooked up
- General restructuring of mod
