+++ build.gradle
@@ -11,7 +11,7 @@ base {
 }
 
 loom {
-    accessWidenerPath = file("src/main/resources/fifym-custom-crafting.accesswidener")
+    accessWidenerPath = file("src/main/resources/fifymcc.accesswidener")
 }
 
 repositories {
