+++ gradle.properties
@@ -10,7 +10,7 @@ org.gradle.jvmargs=-Xmx1G
 # Mod Properties
 	mod_version = 1.0
 	maven_group = net.crioch
-	archives_base_name = fifym-custom-crafting
+	archives_base_name = fifymcc
 
 # Dependencies
 	# check this on https://modmuss50.me/fabric.html
