+++ src/main/java/net/crioch/fifymcc/interfaces/ComponentCraftingInventory.java
@@ -1,0 +1,7 @@
+package net.crioch.fifymcc.interfaces;
+
+import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
+
+public interface ComponentCraftingInventory {
+    default void provideComponentRecipeInputs(ComponentRecipeMatcher finder) {}
+}
