+++ src/main/java/net/crioch/fifymcc/interfaces/ComponentInputSlotFiller.java
@@ -1,0 +1,13 @@
+package net.crioch.fifymcc.interfaces;
+
+import net.minecraft.item.ItemStack;
+import net.minecraft.recipe.RecipeEntry;
+
+import java.util.Iterator;
+
+public interface ComponentInputSlotFiller {
+
+    default void  alignComponentRecipeToGrid(int gridWidth, int gridHeight, int gridOutputSlot, RecipeEntry<?> recipe, Iterator<ItemStack> inputs, int amount) {
+
+    }
+}
