+++ src/main/java/net/crioch/fifymcc/interfaces/ComponentPlayerInventory.java
@@ -1,0 +1,8 @@
+package net.crioch.fifymcc.interfaces;
+
+import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
+
+public interface ComponentPlayerInventory {
+    default void populateComponentRecipeFinder(ComponentRecipeMatcher matcher) {
+    }
+}
