+++ src/main/java/net/crioch/fifymcc/interfaces/ComponentRecipeInputProvider.java
@@ -1,0 +1,8 @@
+package net.crioch.fifymcc.interfaces;
+
+import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
+
+public interface ComponentRecipeInputProvider {
+    default void provideComponentRecipeInputs(ComponentRecipeMatcher finder) {
+    }
+}
