+++ src/main/java/net/crioch/fifymcc/mixin/recipeinputprovider/AbstractFurnaceBlockEntityMixin.java
@@ -1,0 +1,22 @@
+package net.crioch.fifymcc.mixin.recipeinputprovider;
+
+import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
+import net.crioch.fifymcc.interfaces.ComponentRecipeInputProvider;
+import net.minecraft.block.entity.AbstractFurnaceBlockEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.collection.DefaultedList;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+@Mixin(AbstractFurnaceBlockEntity.class)
+public class AbstractFurnaceBlockEntityMixin implements ComponentRecipeInputProvider {
+    @Shadow
+    protected DefaultedList<ItemStack> inventory;
+
+    @Override
+    public void provideComponentRecipeInputs(ComponentRecipeMatcher finder) {
+        for (ItemStack itemStack : this.inventory) {
+            finder.addInput(itemStack);
+        }
+    }
+}
