+++ src/main/java/net/crioch/fifymcc/mixin/screenhandler/PlayerScreenHandlerMixin.java
@@ -1,0 +1,34 @@
+package net.crioch.fifymcc.mixin.screenhandler;
+
+import net.crioch.fifymcc.recipe.ComponentRecipeMatcher;
+import net.crioch.fifymcc.interfaces.ComponentRecipeInputProvider;
+import net.crioch.fifymcc.interfaces.ComponentRecipeScreenHandler;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.inventory.RecipeInputInventory;
+import net.minecraft.recipe.Recipe;
+import net.minecraft.recipe.RecipeEntry;
+import net.minecraft.screen.PlayerScreenHandler;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+@Mixin(PlayerScreenHandler.class)
+public class PlayerScreenHandlerMixin implements ComponentRecipeScreenHandler<RecipeInputInventory> {
+    @Final
+    @Shadow
+    private RecipeInputInventory craftingInput;
+
+    @Final
+    @Shadow
+    private PlayerEntity owner;
+
+    @Override
+    public void populateComponentRecipeFinder(ComponentRecipeMatcher finder) {
+        ((ComponentRecipeInputProvider)this.craftingInput).provideComponentRecipeInputs(finder);
+    }
+
+    @Override
+    public boolean matchesWithComponents(RecipeEntry<? extends Recipe<RecipeInputInventory>> recipe) {
+        return recipe.value().matches(this.craftingInput, this.owner.getWorld());
+    }
+}
