+++ src/main/java/net/crioch/fifymcc/util/Util.java
@@ -1,0 +1,9 @@
+package net.crioch.fifymcc.util;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Util {
+    public static final String MOD_ID = "fifymcc";
+    public static final Logger LOGGER = LoggerFactory.getLogger("FIFYM: Custom Crafting");
+}
