--- src/main/java/net/crioch/fifymcustomcrafting/Util.java
@@ -1,8 +1,0 @@
-package net.crioch.fifymcustomcrafting;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class Util {
-    public static final Logger LOGGER = LoggerFactory.getLogger("fifym-custom-crafting");
-}
