--- src/main/java/net/crioch/fifymcustomcrafting/interfaces/ComponentCraftingInventory.java
@@ -1,7 +1,0 @@
-package net.crioch.fifymcustomcrafting.interfaces;
-
-import net.crioch.fifymcustomcrafting.ComponentRecipeMatcher;
-
-public interface ComponentCraftingInventory {
-    default void provideComponentRecipeInputs(ComponentRecipeMatcher finder) {}
-}
