--- src/main/java/net/crioch/fifymcustomcrafting/interfaces/ComponentPlayerInventory.java
@@ -1,8 +1,0 @@
-package net.crioch.fifymcustomcrafting.interfaces;
-
-import net.crioch.fifymcustomcrafting.ComponentRecipeMatcher;
-
-public interface ComponentPlayerInventory {
-    default void populateComponentRecipeFinder(ComponentRecipeMatcher matcher) {
-    }
-}
