--- src/main/java/net/crioch/fifymcustomcrafting/interfaces/ComponentRecipeInputProvider.java
@@ -1,8 +1,0 @@
-package net.crioch.fifymcustomcrafting.interfaces;
-
-import net.crioch.fifymcustomcrafting.ComponentRecipeMatcher;
-
-public interface ComponentRecipeInputProvider {
-    default void provideComponentRecipeInputs(ComponentRecipeMatcher finder) {
-    }
-}
