+++ src/main/resources/assets/fifymcc/icon.png
