--- src/main/resources/fifym-custom-crafting.client.mixins.json
@@ -1,14 +1,0 @@
-{
-	"required": true,
-	"package": "net.crioch.fifymcustomcrafting.mixin",
-	"compatibilityLevel": "JAVA_21",
-	"client": [],
-	"injectors": {
-		"defaultRequire": 1
-	},
-	"mixins": [
-		"client.RecipeBookWidgetMixin",
-		"client.RecipeResultCollectionMixin",
-		"entity.PlayerInventoryMixin"
-	]
-}
