--- src/main/resources/fifym-custom-crafting.mixins.json
@@ -1,22 +1,0 @@
-{
-  "required": true,
-  "package": "net.crioch.fifymcustomcrafting.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "mixins": [
-    "Ingredient$StackEntryMixin",
-    "IngredientMixin",
-    "InputSlotFillerMixin",
-    "ShapelessRecipeMixin",
-    "entity.PlayerInventoryMixin",
-    "recipeinputprovider.AbstractFurnaceBlockEntityMixin",
-    "recipeinputprovider.CrafterBlockEntityMixin",
-    "recipeinputprovider.CraftingInventoryMixin",
-    "recipeinputprovider.SimpleInventoryMixin",
-    "screenhandler.AbstractFurnaceScreenHandlerMixin",
-    "screenhandler.CraftingScreenHandlerMixin",
-    "screenhandler.PlayerScreenHandlerMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
