+++ src/main/resources/fifymcc.client.mixins.json
@@ -1,0 +1,14 @@
+{
+	"required": true,
+	"package": "net.crioch.fifymcc.mixin",
+	"compatibilityLevel": "JAVA_21",
+	"client": [],
+	"injectors": {
+		"defaultRequire": 1
+	},
+	"mixins": [
+		"client.RecipeBookWidgetMixin",
+		"client.RecipeResultCollectionMixin",
+		"entity.PlayerInventoryMixin"
+	]
+}
